Move usage tracker, operator update tests to their own files
